Problem: main was typed incorrectly
Solution: type main correctly (as integer, not non_neg_integer).

That brings up the concern of Dialyzer not being able to tell between
the two. Actually, it should be inferred that since return type of
`higher_order` is `integer`, the same should be the reutrn type of
`main`! Probably worth investigating / looking up the reason.